requirements.txt
pytest
tabulate
pillow